backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flasgger==0.9.7.1
python-dotenv==1.0.0
Werkzeug==2.3.7
sendgrid==6.10.0
cloudinary==1.36.0
Pillow==10.0.1
psycopg2-binary==2.9.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0